[refactor] Types related to api gathered in one place
- Collect client types in a client folder
- Collect server types in a server folder
