+++ src/@types/types.ts
@@ -1,16 +1,6 @@
 /**
  * temporary folder for type storing
  */
-// UserAuth
-export const RoleType = {
-  ADMIN: 'ADMIN',
-  MANAGER: 'MANAGER',
-  USER: 'USER',
-} as const;
-export type RoleType = typeof RoleType[keyof typeof RoleType];
-
-export const NOTLOGIN = 'NOTLOGIN' as const;
-export type NOTLOGIN = typeof NOTLOGIN;
 
 // NavItem
 export const NavItem = {
