+++ src/components/containers/LicenseListTab/Table/index.tsx
@@ -5,11 +5,8 @@ import useFetch from '@hooks/useFetch';
 import { LicenseListAttr, Number, Trash } from '@@types/types';
 import { licenseListAttr, LIMIT } from '@common/constants';
 import { licenseSearchRequestClient2Server, licenseSearchResponseServer2Client } from '@converter/license';
-import {
-  getLicenseListAPI,
-  GetLicenseListRequestParamsClientType,
-  GetLicenseListResponseClientType,
-} from '@apis/license';
+import { getLicenseListAPI } from '@apis/license';
+import { GetLicenseListRequestParamsClientType, GetLicenseListResponseClientType } from '@@types/client';
 
 export type SearchInfoType = Omit<GetLicenseListRequestParamsClientType, 'limit' | 'offset'>;
 export type ItemType = { [key in LicenseListAttr]: string };
