+++ src/converter/user.ts
@@ -1,20 +1,16 @@
-import { SigninRequestBodyClientType, SigninResponseClientType } from '@apis/user';
-import { RoleServerType, serverRole2ClientRole } from './common';
-import { ReloadResponseClientType, LogoutRequestBodyClientType } from '../apis/user';
-
-// signin
-interface SigninRequestBodyServerType {
-  user_id: string;
-  password: string;
-}
-
-export interface SigninResponseServerType {
-  token_type: string;
-  user_id: string;
-  access_token: string;
-  uuid: string;
-  role: RoleServerType;
-}
+import { serverRole2ClientRole } from './common';
+import {
+  LogoutRequestBodyClientType,
+  ReloadResponseClientType,
+  SigninRequestBodyClientType,
+  SigninResponseClientType,
+} from '@@types/client';
+import {
+  LogoutRequestBodyServerType,
+  ReloadResponseServerType,
+  SigninRequestBodyServerType,
+  SigninResponseServerType,
+} from '@@types/server';
 
 export const signinRequestClient2Server = ({
   id,
@@ -43,25 +39,12 @@ export const signinResponseServer2Client = ({
   };
 };
 
-// reload
-export interface ReloadResponseServerType {
-  user_id: string;
-  role: RoleServerType;
-}
-
 export const reloadResponseServer2Client = ({ user_id, role }: ReloadResponseServerType): ReloadResponseClientType => {
   return {
     id: user_id,
     role: serverRole2ClientRole(role),
   };
 };
-
-// logout
-export interface LogoutRequestBodyServerType {
-  user_id: string;
-  access_token: string;
-  uuid: string;
-}
 
 export const logoutRequestClient2Server = ({
   id,
